+++ extensions/deepwiki/CHANGELOG.md
@@ -3,5 +3,5 @@
 ## [Initial Version] - {PR_MERGE_DATE}
 
 - Initial release of Deepwiki extension.
-- Added command to search repositories on DeepWiki.
-- Added command to open DeepWiki page directly via URL or org/repo identifier.
+- Added command to search repositories on Deepwiki.
+- Added command to open Deepwiki page directly via URL or org/repo identifier.